Added Controllers
Added controllers as well as moved the code to add different features.
