+++ index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const path = require('path');
+const { convertStringToJson } = require('./controllers/convertController');
 
 const app = express();
 const port = 3000;
@@ -17,31 +18,8 @@ app.get('/', (req, res) => {
     res.sendFile(path.join(__dirname, 'public', 'index.html'));
 });
 
-app.post('/convert', (req, res) => {
-    const requestBody = req.body;
-
-    try {
-        // Directly parse the JSON string if it is valid
-        const jsonResult = JSON.parse(JSON.parse(requestBody));
-
-        console.log(jsonResult);
-
-       
-        // Respond with the parsed JSON
-        res.status(200).json({
-            success: true,
-            message: 'String successfully converted to JSON',
-            data: jsonResult,
-        });
-    } catch (error) {
-        // If parsing fails, handle gracefully
-        res.status(400).json({
-            success: false,
-            message: 'Invalid JSON string provided',
-            error: error.message,
-        });
-    }
-});
+// Use the controller for the convert route
+app.post('/convert', convertStringToJson);
 
 // Start the server
 app.listen(port, () => {
